refactor contacts table components
